Update README.md
I'm planning to archive this repo and add a copyleft license. I've received too many emails/pings *expecting* support for this senior design project (StackOverflow, here on GH, and personal/elsewhere), so I plan to put something in place that limits accessibility for for-profit activities with this design. If you're seeing this, feel free to reach out if you have concerns. 

Final branch for your review...